WIP: Pass az_hash and network for the Link object.
We probably shouldn't be going down this path because we're putting a
lot of encoding into the links resolution code. That sort of thing
should be handled *after* we've determined the structure of the
deployment plan.

For AZs:
The tests pass, the code is able to generate a lookup table for the
availability zones, it can resolve the azs in the LinksResolver, and it
can insert the azs into the EvaluationLink object.

For network:
We are inserting the network for the instance into the EvaluationLink,
and if the network isn't specified we fall back to the addressable
network. Not sure that this works for cross-deployment or provided
links. Also not sure that it is sound from an API perspective. Danny
pointed out we should be using the network name that the instances
"hide" behind with DNS.

Signed-off-by: Difan Zhao <dzhao@pivotal.io>